+++ BGL/test/BGL/test_cgal_bgl_named_params.cpp
@@ -91,6 +91,7 @@ void test(const NamedParameters& np)
   assert(get_parameter(np, CGAL::internal_np::use_area_smoothing).v == 54);
   assert(get_parameter(np, CGAL::internal_np::use_Delaunay_flips).v == 55);
   assert(get_parameter(np, CGAL::internal_np::use_safety_constraints).v == 56);
+  assert(get_parameter(np, CGAL::internal_np::dry_run).v == 58);
 
     // Named parameters that we use in the package 'Surface Mesh Simplification'
   assert(get_parameter(np, CGAL::internal_np::get_cost_policy).v == 34);
@@ -175,6 +176,7 @@ void test(const NamedParameters& np)
   check_same_type<54>(get_parameter(np, CGAL::internal_np::use_area_smoothing));
   check_same_type<55>(get_parameter(np, CGAL::internal_np::use_Delaunay_flips));
   check_same_type<56>(get_parameter(np, CGAL::internal_np::use_safety_constraints));
+  check_same_type<58>(get_parameter(np, CGAL::internal_np::dry_run));
 
     // Named parameters that we use in the package 'Surface Mesh Simplification'
   check_same_type<34>(get_parameter(np, CGAL::internal_np::get_cost_policy));
@@ -265,6 +267,7 @@ int main()
                          .use_area_smoothing(A<54>(54))
                          .use_Delaunay_flips(A<55>(55))
                          .use_safety_constraints(A<56>(56))
+                         .dry_run(A<58>(58))
        );
 
   return EXIT_SUCCESS;